Refactored Resets CRUD
